laba_3

+++ app/controllers/user_controller.py
@@ -4,7 +4,7 @@ from litestar.params import Parameter
 from litestar.exceptions import NotFoundException
 from sqlalchemy.ext.asyncio import AsyncSession
 
-from app.User_schem import UserCreate, UserUpdate, UserResponse
+from app.User_schem import UserCreate, UserUpdate, UserResponse, UsersListResponse
 from app.services.user_service import UserService
 
 
@@ -18,9 +18,13 @@ class UserController(Controller):
             session: AsyncSession,
             count: int = Parameter(default=10, gt=0, le=100),
             page: int = Parameter(default=1, gt=0),
-    ) -> List[UserResponse]:
+    ) -> UsersListResponse:
         users = await user_service.get_by_filter(session, count=count, page=page)
-        return [UserResponse.model_validate(user) for user in users]
+        total_count = await user_service.get_total_count(session)
+        return UsersListResponse(
+            users=[UserResponse.model_validate(user) for user in users],
+            total_count=total_count
+        )
 
     @get("/{user_id:str}")
     async def get_user_by_id(
@@ -39,7 +43,7 @@ class UserController(Controller):
             self,
             user_service: UserService,
             session: AsyncSession,
-            data: UserCreate,  # ← ИМЯ "data" для тела запроса
+            data: UserCreate,
     ) -> UserResponse:
         user = await user_service.create(session, data)
         return UserResponse.model_validate(user)
@@ -51,6 +55,9 @@ class UserController(Controller):
             session: AsyncSession,
             user_id: str,
     ) -> None:
+        user = await user_service.get_by_id(session, user_id)
+        if not user:
+            raise NotFoundException(detail=f"User with ID {user_id} not found")
         await user_service.delete(session, user_id)
 
     @put("/{user_id:str}")
@@ -59,7 +66,10 @@ class UserController(Controller):
             user_service: UserService,
             session: AsyncSession,
             user_id: str,
-            data: UserUpdate,  # ← ИМЯ "data"
+            data: UserUpdate,
     ) -> UserResponse:
-        user = await user_service.update(session, user_id, data)
+        try:
+            user = await user_service.update(session, user_id, data)
+        except ValueError as e:
+            raise NotFoundException(detail=str(e))
         return UserResponse.model_validate(user)
